+++ app/Jobs/Issues/CreateNewReminder.php
@@ -38,6 +38,9 @@ class CreateNewReminder implements ShouldQueue
      */
     public function handle()
     {
+        if (!is_null($this->issue->timeClosed)) {
+            return;
+        }
         if ($this->typeOfReminder == 'paused') {
             if ($this->issue->paused) {
                 event(new IssuePaused($this->issue, 'paused'));